Load game dialog via menu bar in new desktop client
