+++ desktop/src/app/newgame/newgame.component.html
@@ -1,3 +1,6 @@
+<!-- Show the header on the start page -->
+<app-header></app-header>
+
 <!-- Create a jumbotron component with new game -->
 <div class="jumbotron text-center top-space">
     <h1 class="display-4 text-center"><img src="assets/trams-logo.png" alt="TraMS" class="img-responsive" style="max-height:80px;"></h1>
